fix(Views v2) allow filtering ff link use
