+++ src/Tribe/Views/V2/Messages.php
@@ -94,9 +94,8 @@ class Messages {
 				'There were no results found.',
 				'the-events-calendar'
 			),
-			// translators: the placeholder is an html link to the next month with available events.
 			'month_no_results_found'           => __(
-				'There were no results found for this view. %1$s',
+				'There were no results found for this view.',
 				'the-events-calendar'
 			),
 			// translators: the placeholder is the keyword(s), as the user entered it in the bar.
@@ -109,8 +108,18 @@ class Messages {
 				'There were no results found for <strong>"%1$s"</strong> this month. Try searching next month.',
 				'the-events-calendar'
 			),
-			// translators: %1$s: events (plural), %2$s: the formatted date string, e.g. "February 22, 2020". %3$s html link to next day with available events.
+			// translators: %1$s: events (plural), %2$s: the formatted date string, e.g. "February 22, 2020".
 			'day_no_results_found'             => __(
+				'No %1$s scheduled for %2$s.',
+				'the-events-calendar'
+			),
+			// translators: the placeholder is an html link to the next month with available events.
+			'month_no_results_found_w_ff_link'           => __(
+				'There were no results found for this view. %1$s',
+				'the-events-calendar'
+			),
+			// translators: %1$s: events (plural), %2$s: the formatted date string, e.g. "February 22, 2020". %3$s html link to next day with available events.
+			'day_no_results_found_w_ff_link'             => __(
 				'No %1$s scheduled for %2$s. %3$s',
 				'the-events-calendar'
 			),
